TST: Reorganize model testing functions into different scripts
Reorganize model testing functions into different scripts:
- Move dMRI model testing into a dedicated script. Keep in the existing
  testing module tests related to the base module, including those using
  the model factory, as it can instantiate both DWI and PET models.
- Rename the script testing the GPR module to increase consistency with
  the current reorganization: add the `model` label to the filename.

Splitting the tests corresponding to the `model.dmri` module:
- Follows the approach adopted for the data classes.
- Improves compartmentalization.
- Potentially removes unrelated import statements, definitions, etc.
  from each module, improving readability.
- Allows to more easily locate tests corresponding to a given module.
- Potentially allows identifying more easily missing tests (e.g. testing
  the model factory with particular models or datasets).

Adding the `model` label to the script testing the GPR module allows to
more easily identify its purpose.
