+++ test/test_model_dmri.py
@@ -1,0 +1,208 @@
+# emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
+# vi: set ft=python sts=4 ts=4 sw=4 et:
+#
+# Copyright The NiPreps Developers <nipreps@gmail.com>
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# We support and encourage derived works from this project, please read
+# about our expectations at
+#
+#     https://www.nipreps.org/community/licensing/
+#
+"""Unit tests exercising dMRI models."""
+
+import warnings
+
+import numpy as np
+import pytest
+from dipy.sims.voxel import single_tensor
+
+from nifreeze import model
+from nifreeze.data.dmri import DWI
+from nifreeze.data.dmri.base import DWI_REDUNDANT_B0_WARN_MSG
+from nifreeze.data.dmri.utils import (
+    DEFAULT_LOWB_THRESHOLD,
+    DTI_MIN_ORIENTATIONS,
+)
+from nifreeze.model._dipy import GaussianProcessModel
+from nifreeze.testing import simulations as _sim
+
+B_MATRIX = np.array(
+    [
+        [0.0, 0.0, 0.0, 0],
+        [1.0, 0.0, 0.0, 500],
+        [0.0, 1.0, 0.0, 1000],
+        [0.0, 0.0, 1.0, 1000],
+        [1 / np.sqrt(2), 1 / np.sqrt(2), 0.0, 1000],
+        [1 / np.sqrt(2), 0.0, 1 / np.sqrt(2), 1000],
+        [0.0, 1 / np.sqrt(2), 1 / np.sqrt(2), 1000],
+        [1 / np.sqrt(3), 1 / np.sqrt(3), 1 / np.sqrt(3), 2000],
+        [-1 / np.sqrt(3), 1 / np.sqrt(3), 1 / np.sqrt(3), 2000],
+        [1 / np.sqrt(3), -1 / np.sqrt(3), 1 / np.sqrt(3), 2000],
+        [1 / np.sqrt(3), 1 / np.sqrt(3), -1 / np.sqrt(3), 2000],
+    ],
+    dtype=np.float32,
+)
+
+
+def test_average_model():
+    """Check the implementation of the average DW model."""
+
+    gtab = B_MATRIX.copy()
+    size = (10, 10, 10, gtab.shape[0])
+    data = np.ones(size, dtype=float)
+    mask = np.ones(size[:3], dtype=bool)
+
+    data *= gtab[:, -1]
+    dataset = DWI(dataobj=data, affine=np.eye(4), gradients=gtab, brainmask=mask)
+
+    avgmodel_mean = model.AverageDWIModel(dataset, stat="mean")
+    avgmodel_mean_full = model.AverageDWIModel(dataset, stat="mean", atol_low=2000, atol_high=2000)
+    avgmodel_median = model.AverageDWIModel(dataset)
+
+    # Verify that average cannot be calculated in shells with one single value
+    # The two first gradients are considered low-b orientations because of the
+    # default threshold and are pulled out (so now index is 0 for b=500).
+    with pytest.raises(RuntimeError):
+        avgmodel_mean.fit_predict(0)
+
+    assert np.allclose(avgmodel_mean.fit_predict(3), 1000)
+    assert np.allclose(avgmodel_median.fit_predict(3), 1000)
+
+    grads = list(gtab[1:, -1])  # Exclude b0
+    del grads[3]
+    assert np.allclose(avgmodel_mean_full.fit_predict(3), np.mean(grads))
+
+    avgmodel_mean_2000 = model.AverageDWIModel(dataset, stat="mean", atol_low=1100)
+    avgmodel_median_2000 = model.AverageDWIModel(dataset, atol_low=1100)
+
+    last = gtab.shape[0] - 2  # Last but one index (b=2000)
+    assert np.allclose(avgmodel_mean_2000.fit_predict(last), gtab[2:-1, -1].mean())
+    assert np.allclose(avgmodel_median_2000.fit_predict(last), 1000)
+
+
+@pytest.mark.parametrize(
+    (
+        "bval_shell",
+        "S0",
+        "evals",
+    ),
+    [
+        (
+            1000,
+            100,
+            (0.0015, 0.0003, 0.0003),
+        )
+    ],
+)
+@pytest.mark.parametrize("snr", (10, 20))
+@pytest.mark.parametrize("hsph_dirs", (60, 30))
+def test_gp_model(evals, S0, snr, hsph_dirs, bval_shell):
+    # Simulate signal for a single tensor
+    evecs = _sim.create_single_fiber_evecs()
+    gtab = _sim.create_single_shell_gradient_table(hsph_dirs, bval_shell)
+    signal = single_tensor(gtab, S0=S0, evals=evals, evecs=evecs, snr=snr)
+
+    # Drop the initial b=0
+    gtab = gtab[1:]
+    data = signal[1:]
+
+    gp = GaussianProcessModel(kernel_model="spherical")
+    assert isinstance(gp, model._dipy.GaussianProcessModel)
+
+    gpfit = gp.fit(data[:-2], gtab[:-2])
+    prediction = gpfit.predict(gtab.bvecs[-2:])
+
+    assert prediction.shape == (2,)
+
+
+@pytest.mark.random_dwi_data(50, (14, 16, 8), True)
+def test_dti_model(setup_random_dwi_data):
+    (
+        dwi_dataobj,
+        affine,
+        brainmask_dataobj,
+        b0_dataobj,
+        gradients,
+        _,
+    ) = setup_random_dwi_data
+
+    # Ignore warning due to redundant b0 volumes
+    with warnings.catch_warnings():
+        warnings.filterwarnings("ignore", message=DWI_REDUNDANT_B0_WARN_MSG, category=UserWarning)
+        dataset = DWI(
+            dataobj=dwi_dataobj,
+            affine=affine,
+            brainmask=brainmask_dataobj,
+            bzero=b0_dataobj,
+            gradients=gradients,
+        )
+
+    dtimodel = model.DTIModel(dataset)
+    predicted = dtimodel.fit_predict(4)
+    assert predicted is not None
+    assert predicted.shape == dwi_dataobj.shape[:-1]
+
+
+def test_dmri_exceptions():
+    import re
+
+    class DummyDWI(DWI):
+        def __init__(
+            self, bzero=True, gradients=True, data_shape=(10, 10, 10, 35), brainmask=None
+        ):
+            if bzero:
+                self.bzero = np.ones(data_shape[:3])
+            if gradients:
+                self.gradients = np.ones((data_shape[-1], 4))
+            self.dataobj = np.ones(data_shape)
+            self.brainmask = brainmask
+
+        def __len__(self):
+            if hasattr(self, "gradients") and self.gradients is not None:
+                return self.gradients.shape[0]
+            return 0
+
+    dwi = DummyDWI(bzero=False)
+    with pytest.raises(TypeError, match=model.dmri.DWI_OBJECT_ERROR_MSG):
+        model.dmri.BaseDWIModel(dwi)
+
+    dwi = DummyDWI(gradients=False)
+    with pytest.raises(ValueError, match=model.dmri.DWI_GTAB_ERROR_MSG):
+        model.dmri.BaseDWIModel(dwi)
+
+    min_dir = DTI_MIN_ORIENTATIONS - 1
+    dwi = DummyDWI(data_shape=(10, 10, 10, min_dir))
+    with pytest.raises(
+        ValueError, match=re.escape(model.dmri.DWI_SIZE_ERROR_MSG.format(directions=min_dir))
+    ):
+        model.dmri.BaseDWIModel(dwi)
+
+
+def test_dmri_max_b_attribute():
+    class DummyDWI(DWI):
+        def __init__(self, data_shape=(10, 10, 10, 35)):
+            self.bzero = np.ones(data_shape[:3])
+            self.gradients = np.ones((data_shape[-1], 4))
+            self.dataobj = np.ones(data_shape)
+            self.brainmask = np.ones(data_shape[:-1]).astype(bool)
+
+        def __len__(self):
+            return self.gradients.shape[0]
+
+    dwi = DummyDWI()
+    max_b = DEFAULT_LOWB_THRESHOLD + 10
+    dwi_base_model = model.dmri.BaseDWIModel(dwi, max_b=max_b)
+    assert hasattr(dwi_base_model, "_max_b")
+    assert dwi_base_model._max_b == max_b
